requirements.txt
aiogram>=3.0,<4.0
redis
motor
python-dotenv
pyyaml
